Update project layout
