+++ pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "wakeword_collector"
+version = "0.0.1"
+dynamic = ["dependencies"]
+
+[tool.setuptools.dynamic]
+dependencies = {file = ["requirements.txt"]}
+
+[build-system]
+build-backend = "setuptools.build_meta"
+requires = [
+    "setuptools ~=63.2.0",
+    "wheel ~=0.37.1",
+]
+
+[project.scripts]
+wakeword_collect = "wakeword_collector.__main__:main"
+
+
