+++ requirements.txt
@@ -6,5 +6,5 @@ PyAudio==0.2.11
 pylint==2.6.0
 six==1.15.0
 toml==0.10.2
-typed-ast==1.4.1
+typed-ast
 wrapt==1.12.1
